+++ Api-Antivirus/Interface/IInstitutions.cs
@@ -5,9 +5,9 @@ namespace Api_Antivirus.Interface
     public interface IInstitutions
     {
         Task<IEnumerable<InstitutionsResponseDto>> GetAllAsync();
-        Task<InstitutionsResponseDto> GetByIdAsync(long id);
+        Task<InstitutionsResponseDto> GetByIdAsync(int id);
         Task CreateAsync(InstitutionsRequestDto dto);
-        Task UpdateAsync(long id, InstitutionsRequestDto dto);
-        Task DeleteAsync(long id);
+        Task UpdateAsync(int id, InstitutionsRequestDto dto);
+        Task DeleteAsync(int id);
     }
 }
