+++ Api-Antivirus/Services/UsersService.cs
@@ -28,7 +28,7 @@ namespace Api_Antivirus.Services
             _mapper.Map<UsersRequestDto>(entity);
         }
 
-        public async Task DeleteAsync(long id)
+        public async Task DeleteAsync(int id)
         {
             var entity = await _context.users.FindAsync(id);
             if (entity != null)
@@ -44,13 +44,13 @@ namespace Api_Antivirus.Services
             return _mapper.Map<IEnumerable<UsersResponseDto>>(entities);
         }
 
-        public async Task<UsersResponseDto> GetByIdAsync(long id)
+        public async Task<UsersResponseDto> GetByIdAsync(int id)
         {
             var entity = await _context.users.FindAsync(id);
             return _mapper.Map<UsersResponseDto>(entity);
         }
 
-        public async Task UpdateAsync(long id, UsersRequestDto dto)
+        public async Task UpdateAsync(int id, UsersRequestDto dto)
         {
             var entity = await _context.users.FindAsync(id);
             if (entity != null)
